replace render when hideFormatting is true
